Generate dynamic library file for Python bindings
Using PyO3 it is possible to compile the library in a dynamic library
format (such as *.so on Linux). This generated dynamic library can
then be imported via other via other programming languages such as
Python and C.

On running:
```shell
$ cargo build
```
A dynamic library file will be generated in `target/debug/`
with the name, such as on Linux: `libauto_image_cropper.so`.

By changing the current directory to `target/debug/`, this
can now be imported and corresponding functions may be called, in
say Python with:
```python
>>> import libauto_image_cropper
>>> libauto_image_cropper.calculate_corners("/path/to/imagefile.png")
```